broadcasting gamestate TODO: separate Players and Network clients networkmanager should only deal with clients

+++ model/player.go
@@ -27,7 +27,7 @@ type Controls struct {
 
 // Player ...
 type Player struct {
-	ID         int
+	ID         uint8
 	Control    Controls
 	Collider   RectCollider
 	NetworkOut chan []byte
